Create README.md
Our top priority is the product delivery in the best possible way. At BeeWeb, everyone shares these beliefs. So just like a bee, we work diligently and try to do our utmost to satisfy our clients’ needs and requirements. We appreciate good, long-term relationships with clients, and serve an individual approach to each client and project.